requirements.txt
requests==2.28.1
streamlit==1.13.0
streamlit_lottie==0.0.2
streamlit-authenticator==0.1.5
streamlit_option_menu==0.3.2
yfinance==0.1.74
pandas==1.3.5
prophet==1.1.1
plotly==5.10.0
tzdata==2021.5
tzlocal==4.2
pytz==2021.3
click==8
deta==1.1.0
python-dotenv==0.21.0